b1_9 hecho, ejerc final comenzado

+++ BLOQUE_01/UD_1_2/b1_2_calculadora_simple.py
@@ -2,13 +2,13 @@
 #(Tb evito que el programa falle)
 #  
 try:
-    num1 = float(input("Ingresa el primer número: "))
-    num2 = float(input("Ingresa el segundo número: "))
+    num1 = int(input("Ingresa el primer número: "))
+    num2 = int(input("Ingresa el segundo número: "))
 
     print(f"Suma: {num1} + {num2} = {num1 + num2}")
     print(f"Resta: {num1} - {num2} = {num1 - num2}")
     print(f"Multiplicación: {num1} * {num2} = {num1 * num2}")
-    
+
     #Aquí saco error si el numero por el que se va a dividir es cero (div y modulo)
     if num2 != 0:
         print(f"División: {num1} / {num2} = {num1 / num2}")
@@ -16,7 +16,7 @@ try:
     else:
         print("División: Error, no se puede dividir entre cero")
         print("Módulo: Error, no se puede calcular el módulo con divisor cero")
-    
+
     print(f"Potencia: {num1} ** {num2} = {num1 ** num2}")
 
 except ValueError:
